Reorganizes source files
The code isn't split up yet, so there is little point in separating
that into pieces, but the rest of it can go ahead and get some.

Specifically, the icons now are in their own areas, the sounds are
split up a bit better. The code is in its own big code folder, to
be split up as it's being rewritten.

Extra things that I have no idea what they are for or don't think
they really work in the 'source' code, are in the extra folder.

The source should be standalone in its own cute little src folder now.

Also slightly cleans up the Login file.

Fixes #1
Fixes #4